// libs/shared/modules/src/risk/models/risk.ts
export interface Risk {
  /**
   * risk id
   */
  id?: string;
  /**
   * risk name
   */
  name?: string;
  /**
   * risk category id
   */
  category_id?: string;
  /**
   * list of risk effects
   */
  effect?: Array<EffectEnum>;
  risk_description?: string;
  threat_description?: string;
  vulnerability_description?: string;
  /**
   * risk source id
   */
  source_id?: string;
  inherent_risk_level?: InherentRiskLevelEnum;
  inherent_risk_level_impact?: InherentRiskLevelImpactEnum;
  inherent_risk_level_like_hood?: InherentRiskLevelLikeHoodEnum;
  inherent_risk_level_financial_impact?: string;
  residual_risk_level_impact?: ResidualRiskLevelImpactEnum;
  residual_risk_level_like_hood?: ResidualRiskLevelLikeHoodEnum;
  residual_risk_level?: ResidualRiskLevelEnum;
  residual_risk_level_financial_impact?: string;
  strategy?: StrategyEnum;
  level_target?: LevelTargetEnum;
  /**
   * list of control ids that related to risk management
   */
  mitigation_control_ids?: Array<string>;
  /**
   * list of evidence ids that related to risk
   */
  evidence_ids?: Array<string>;
  /**
   * The email of control's owner
   */
  owner?: string;
  creator_name?: string;
  creation_time?: string;
  last_updated?: string;
  last_updated_by?: string;
}

export interface DetailedRisk extends Risk {
  isCustomCategory?: boolean;
  isCustomSource?: boolean;
  editedField?: string;
  previousValue?: string;
  newValue?: string;
}

export type EffectEnum = 'Confidentiality' | 'Integrity' | 'Availability';
export const EffectEnum = {
  Confidentiality: 'Confidentiality' as EffectEnum,
  Integrity: 'Integrity' as EffectEnum,
  Availability: 'Availability' as EffectEnum,
};
export type InherentRiskLevelEnum = 'Low' | 'Medium' | 'High' | 'Critical' | '' | 'null';
export const InherentRiskLevelEnum = {
  Low: 'Low' as InherentRiskLevelEnum,
  Medium: 'Medium' as InherentRiskLevelEnum,
  High: 'High' as InherentRiskLevelEnum,
  Critical: 'Critical' as InherentRiskLevelEnum,
  Empty: '' as InherentRiskLevelEnum,
  Null: 'null' as InherentRiskLevelEnum,
};
export type InherentRiskLevelImpactEnum =
  | 'Insignificant'
  | 'Minor'
  | 'Moderate'
  | 'Major'
  | 'Catastrophic'
  | ''
  | 'null';
export const InherentRiskLevelImpactEnum = {
  Insignificant: 'Insignificant' as InherentRiskLevelImpactEnum,
  Minor: 'Minor' as InherentRiskLevelImpactEnum,
  Moderate: 'Moderate' as InherentRiskLevelImpactEnum,
  Major: 'Major' as InherentRiskLevelImpactEnum,
  Catastrophic: 'Catastrophic' as InherentRiskLevelImpactEnum,
  Empty: '' as InherentRiskLevelImpactEnum,
  Null: 'null' as InherentRiskLevelImpactEnum,
};
export type InherentRiskLevelLikeHoodEnum =
  | 'Rare'
  | 'Unlikely'
  | 'Possible'
  | 'Likely'
  | 'Almost Certain'
  | ''
  | 'null';
export const InherentRiskLevelLikeHoodEnum = {
  Rare: 'Rare' as InherentRiskLevelLikeHoodEnum,
  Unlikely: 'Unlikely' as InherentRiskLevelLikeHoodEnum,
  Possible: 'Possible' as InherentRiskLevelLikeHoodEnum,
  Likely: 'Likely' as InherentRiskLevelLikeHoodEnum,
  AlmostCertain: 'Almost Certain' as InherentRiskLevelLikeHoodEnum,
  Empty: '' as InherentRiskLevelLikeHoodEnum,
  Null: 'null' as InherentRiskLevelLikeHoodEnum,
};
export type ResidualRiskLevelImpactEnum =
  | 'Insignificant'
  | 'Minor'
  | 'Moderate'
  | 'Major'
  | 'Catastrophic'
  | ''
  | 'null';
export const ResidualRiskLevelImpactEnum = {
  Insignificant: 'Insignificant' as ResidualRiskLevelImpactEnum,
  Minor: 'Minor' as ResidualRiskLevelImpactEnum,
  Moderate: 'Moderate' as ResidualRiskLevelImpactEnum,
  Major: 'Major' as ResidualRiskLevelImpactEnum,
  Catastrophic: 'Catastrophic' as ResidualRiskLevelImpactEnum,
  Empty: '' as ResidualRiskLevelImpactEnum,
  Null: 'null' as ResidualRiskLevelImpactEnum,
};
export type ResidualRiskLevelLikeHoodEnum =
  | 'Rare'
  | 'Unlikely'
  | 'Possible'
  | 'Likely'
  | 'Almost Certain'
  | ''
  | 'null';
export const ResidualRiskLevelLikeHoodEnum = {
  Rare: 'Rare' as ResidualRiskLevelLikeHoodEnum,
  Unlikely: 'Unlikely' as ResidualRiskLevelLikeHoodEnum,
  Possible: 'Possible' as ResidualRiskLevelLikeHoodEnum,
  Likely: 'Likely' as ResidualRiskLevelLikeHoodEnum,
  AlmostCertain: 'Almost Certain' as ResidualRiskLevelLikeHoodEnum,
  Empty: '' as ResidualRiskLevelLikeHoodEnum,
  Null: 'null' as ResidualRiskLevelLikeHoodEnum,
};
export type ResidualRiskLevelEnum = 'Low' | 'Medium' | 'High' | 'Critical' | '' | 'null' | 'Not Set';
export const ResidualRiskLevelEnum = {
  Low: 'Low' as ResidualRiskLevelEnum,
  Medium: 'Medium' as ResidualRiskLevelEnum,
  High: 'High' as ResidualRiskLevelEnum,
  Critical: 'Critical' as ResidualRiskLevelEnum,
  Empty: '' as ResidualRiskLevelEnum,
  Null: 'null' as ResidualRiskLevelEnum,
  NotSet: 'Not Set' as ResidualRiskLevelEnum,
};
export type StrategyEnum = 'Mitigate' | 'Accept' | 'Transfer' | 'Avoid' | '' | 'null';
export const StrategyEnum = {
  Mitigate: 'Mitigate' as StrategyEnum,
  Accept: 'Accept' as StrategyEnum,
  Transfer: 'Transfer' as StrategyEnum,
  Avoid: 'Avoid' as StrategyEnum,
  Empty: '' as StrategyEnum,
  Null: 'null' as StrategyEnum,
};
export type LevelTargetEnum = 'Low' | 'Medium' | 'High' | 'Critical' | '' | 'null';
export const LevelTargetEnum = {
  Low: 'Low' as LevelTargetEnum,
  Medium: 'Medium' as LevelTargetEnum,
  High: 'High' as LevelTargetEnum,
  Critical: 'Critical' as LevelTargetEnum,
  Empty: '' as LevelTargetEnum,
  Null: 'null' as LevelTargetEnum,
};