// libs/shared/modules/src/risk/models/riskCategory.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface RiskCategory {
  /**
   * risk category name
   */
  category_name?: string;
  /**
   * risk category id
   */
  id?: string;
}