// libs/shared/modules/src/risk/models/riskSource.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface RiskSource { 
    /**
     * risk source name
     */
    source_name?: string;
    /**
     * risk source id
     */
    id?: string;
}